Revert "feat: :sparkles: New !details Command!"
